Module Aggregation.Main_protocol

module SMap = Kzg.SMap

module type S = sig ... end

module Make_impl (Super_PP : Polynomial_protocol.S) : sig ... end

module Make (PP : Polynomial_protocol.S) : S with module PP = PP and type public_inputs = Kzg.Bls.Scalar.t array list

include sig ... end

module PP : sig ... end

exception Rest_not_null of string

exception Entry_not_in_table of string

module Input_commitment : sig ... end

type scalar = Kzg.Bls.Scalar.t

val scalar_t : scalar Repr.t

val scalar_encoding : scalar Data_encoding.t

type circuit_map = (Plonk.Circuit.t * int) Kzg.SMap.t

type prover_public_parameters = Make(Polynomial_protocol).prover_public_parameters

val prover_public_parameters_t : prover_public_parameters Repr.t

type verifier_public_parameters = Make(Polynomial_protocol).verifier_public_parameters

val verifier_public_parameters_t : verifier_public_parameters Repr.t

val verifier_public_parameters_encoding : verifier_public_parameters Data_encoding.t

type proof = Make(Polynomial_protocol).proof

val proof_t : proof Repr.t

val proof_encoding : proof Data_encoding.t

type circuit_prover_input = Make(Polynomial_protocol).circuit_prover_input = {

witness : scalar array;

input_commitments : Input_commitment.t list;

}

val circuit_prover_input_t : circuit_prover_input Repr.t

type prover_inputs = circuit_prover_input list Kzg.SMap.t

val prover_inputs_t : prover_inputs Repr.t

type public_inputs = Kzg.Bls.Scalar.t array list

val public_inputs_t : public_inputs Repr.t

type circuit_verifier_input = Make(Polynomial_protocol).circuit_verifier_input = {

nb_proofs : int;

public : public_inputs;

commitments : Input_commitment.public list list;

}

val circuit_verifier_input_t : circuit_verifier_input Repr.t

type verifier_inputs = circuit_verifier_input Kzg.SMap.t

val verifier_inputs_t : verifier_inputs Repr.t

val to_verifier_inputs : prover_public_parameters -> prover_inputs -> verifier_inputs

val input_commit : ?size:int -> ?shift:int -> prover_public_parameters -> scalar array -> Input_commitment.t

val setup : zero_knowledge:bool -> circuit_map -> srs:(Kzg.Bls.Srs.t * Kzg.Bls.Srs.t) -> prover_public_parameters * verifier_public_parameters

val update_prover_public_parameters : 'a Repr.ty -> 'a -> prover_public_parameters -> prover_public_parameters

val update_verifier_public_parameters : 'a Repr.ty -> 'a -> verifier_public_parameters -> verifier_public_parameters

val prove : prover_public_parameters -> inputs:prover_inputs -> proof

val verify : verifier_public_parameters -> inputs:verifier_inputs -> proof -> bool

module Internal_for_tests : sig ... end

module Gates : sig ... end

module Perm : sig ... end

module RangeCheck : sig ... end

val get_gen_n_prover : prover_public_parameters -> scalar * int

Returns (g, n), where n is the size of the circuit padded to the next power of two & g is a primitive n-th root of unity

val get_gen_n_verifier : verifier_public_parameters -> scalar * int

Returns (g, n), where n is the size of the circuit padded to the next power of two & g is a primitive n-th root of unity

val filter_prv_pp_circuits : prover_public_parameters -> 'a SMap.t -> prover_public_parameters

val filter_vrf_pp_circuits : verifier_public_parameters -> 'a SMap.t -> verifier_public_parameters

type prover_aux = Make(Polynomial_protocol).prover_aux = {

answers : scalar SMap.t SMap.t list;

batch : scalar SMap.t list;

alpha : scalar;

beta : scalar;

gamma : scalar;

delta : scalar;

x : scalar;

r : scalar;

cms_answers : PP.Answers_commitment.t SMap.t;

cms_pi : PP.Answers_commitment.t SMap.t;

ids_batch : (scalar * int) SMap.t;

t_answers : scalar list;

}

Auxiliary information needed by the prover for the meta-verification in aPlonK

type verifier_aux = Make(Polynomial_protocol).verifier_aux = {

alpha : scalar;

beta : scalar;

gamma : scalar;

delta : scalar;

x : scalar;

r : scalar;

}

Auxiliary information needed by the verifier for the meta-verification in aPlonK

type input_commit_funcs = Make(Polynomial_protocol).input_commit_funcs = {

pi : scalar array -> PP.Answers_commitment.t;

answers : Kzg.Bls.Scalar.t SMap.t SMap.t list -> PP.Answers_commitment.t;

}

val prove_list : prover_public_parameters -> input_commit_funcs:input_commit_funcs SMap.t -> inputs:prover_inputs -> proof * prover_aux

val verify_list : verifier_public_parameters -> (proof * scalar SMap.t list * PP.Answers_commitment.public SMap.t * PP.Answers_commitment.public SMap.t * scalar list * (scalar * int) SMap.t) -> bool * verifier_aux