Module Bls12_381_hash.Permutation

module Poseidon : sig ... end

Implementation of Poseidon over the scalar field of BLS12-381 for a security with the permutation x^5.

module Rescue : sig ... end

Implementation of an instantiation of Rescue over the scalar field of BLS12-381 for a security of 128 bits and with the permutation x^5. The parameters of the instantiation are:

module Anemoi : sig ... end

Implementation of the permutation Anemoi and the mode of operation Jive over the scalar field of BLS12-381.

module Griffin : sig ... end

Griffin over the scalar field of BLS12-381 for a security of 128 bits and with the permutation x^5.