Module Poseidon.Parameters

type t = {

state_size : int;

nb_of_partial_rounds : int;

nb_of_full_rounds : int;

batch_size : int;

round_constants : Bls12_381.Fr.t array;

linear_layer : Bls12_381.Fr.t array array;

}

val security_128_state_size_3 : t

Parameters for Poseidon with a state size of 3 for a security of 128bits. FIXME: The linear layer and the round constants are not standard

val security_256_state_size_5 : t

Parameters for Poseidon with a state size of 5 for a security of 256bits. FIXME: The linear layer and the round constants are not standard