Module type Bls12_381_hash.PERMUTATION

type parameters

Parameters for a specific instance

type ctxt

Context of the permutation

val allocate_ctxt : parameters -> ctxt

allocate_ctxt parameters. Allocate a context for a specific instance of the permutation.

val get_state : ctxt -> Bls12_381.Fr.t array

Return the current state of the context

val get_state_size : ctxt -> int

Return the state size of the context

val set_state : ctxt -> Bls12_381.Fr.t array -> unit

set_state ctxt state. Set the context state to the given value. The value state must be of the same size than the expecting state

val apply_permutation : ctxt -> unit

Apply a permutation on the current state of the context