Module MinPk.Aug

Augmentation scheme described in section 3.2

In a message augmentation scheme, signatures are generated over the concatenation of the public key and the message, ensuring that messages signed by different public keys are distinct.

val sign : sk -> Stdlib.Bytes.t -> signature

sign sk msg implements the algorithm described in section 3.2.1

val verify : pk -> Stdlib.Bytes.t -> signature -> bool

verify pk msg signature implements the algorithm described in section 3.2.2

val aggregate_verify : (pk * Stdlib.Bytes.t) list -> signature -> bool

aggregate_verify pks msg aggregated_signature performs a aggregate signature verification. It implements the AggregateVerify algorithm specified in section 3.2.3