Module Io_array.Make

Takes an IO instance and wraps it in an Array interface with support for prefetching sections of the array.

Parameters

module IO : Io.S

module Elt : ELT

Signature

include Search.ARRAY with type elt = Elt.t

type t

type elt = Elt.t

val get : t -> Import.int63 -> elt

val length : t -> Import.int63

val pre_fetch : t -> low:Import.int63 -> high:Import.int63 -> unit

type io = IO.t

val v : io -> t