Module Brassaia_pack_unix.Pack_value

include module type of struct include Brassaia_pack.Pack_value end

module Kind = Brassaia_pack.Pack_value.Kind

type nonrec weight =

| Immediate of int

| Deferred of unit -> int

type nonrec kinded = ..

kinded is an extenisble variant that each S extends so that it can define S.to_kinded and S.of_kinded. Its purpose is to allow containers, such as Brassaia_pack_unix.Lru, to store and return all types of S and thus be usable by modules defined over S, such as Brassaia_pack_unix.Pack_store.

module type S = Brassaia_pack.Pack_value.S

module type Config = Brassaia_pack.Pack_value.Config

module Of_contents = Brassaia_pack.Pack_value.Of_contents

module Of_commit = Brassaia_pack.Pack_value.Of_commit

module type Persistent = sig ... end