Module D.Remote_config

The configuration of a node to be run as a remote node i.e., one that can both send an receive messages with other nodes.

type t = {

remote_nodes : (string * int * string) list;
(*
The initial list of remote nodes which this node can send messages to. A list of external ip address/port/node name triplets.
*)

local_port : int;
(*
The port that this node should run on.
*)

connection_backlog : int;
(*
The the argument used when listening on a socket.
*)

node_name : string;
(*
The name of this node.
*)

node_ip : string;
(*
The external ip address of this node.
*)

}