Module Distribution.Polynomial_commitment

module SMap = Kzg.SMap

Extension of the PC signature with additional types and functions used in by Distributed_prover

module type PC_for_distribution_sig = sig ... end

module Kzg_impl : sig ... end

Extension of the KZG implementation with additional types and functions used in by Distributed_prover