Module PBT.MakeEquality

Parameters

module G : sig ... end

Signature

val zero : unit -> unit

Verify the equality is correct with the value zero

val one : unit -> unit

Verify the equality is correct with the value one

val random_same_objects : unit -> unit

Verify the equality of two random values created invidually

val get_tests : unit -> string * unit Alcotest.test_case list

Returns the tests to be used with Alcotest