Module type CurveSig.PAIRING

module G1 : BASE

module G2 : BASE

module GT : Bls12_381.Ff_sig.BASE

exception FailToComputeFinalExponentiation of GT.t

val miller_loop : (G1.t * G2.t) list -> GT.t

val miller_loop_simple : G1.t -> G2.t -> GT.t

Compute the miller loop on a single tuple of point

val pairing : G1.t -> G2.t -> GT.t

Compute a pairing result of a list of points

val final_exponentiation_opt : GT.t -> GT.t option

Compute the final exponentiation of the given point. Returns a None if the point is null

val final_exponentiation_exn : GT.t -> GT.t

Compute the final exponentiation of the given point. Raise FailToComputeFinalExponentiation if the point is null