Module Make.Hash

type scalar = Scalar.t

type ctxt

val init : ?input_length:int -> unit -> ctxt

Initialize a raw hash context

val digest : ctxt -> scalar array -> ctxt

digest ctxt input computes the hash of the given input

val get : ctxt -> scalar

get ctxt returns the resulting point after hash has been called