Module Sinsemilla.Zcash

val hash_exn : Mec__.Pallas.Affine.t -> Mec__.Iterator.Bit.t -> Mec__.Pallas.Affine.Base.t

hash_exn initial_value iterator is the instantiation of Sinsemilla with the Zcash generators and a chunk size of 10

val hash_opt : Mec__.Pallas.Affine.t -> Mec__.Iterator.Bit.t -> Mec__.Pallas.Affine.Base.t option

hash_opt initial_value iterator is the instantiation of Sinsemilla with the Zcash generators and a chunk size of 10