Module Permutation.Hades

HADES Strategy, based on On a Generalization of Substitution-Permutation Networks: The HADES Design Strategy

module type PARAMETERS = sig ... end

module Make (Param : PARAMETERS) (Scalar : Bls12_381.Ff_sig.PRIME) : sig ... end

Build an HADES SP network based on the parameters and a scalar field