Module type Hades.PARAMETERS

val width : int

The state size

val full_rounds : int

The total number of full rounds

val partial_rounds : int

The number of partial rounds

val round_constants : string array

The round constants, given in decimal representation

Secure round constants can be constructed using Sage scripts provided in this repository

val linear_transformation : string array array

The linear transformation, given in decimal representation.

Secure linear transformations can be constructed using Sage scripts provided in this repository

val partial_round_idx_to_permute : int

The index of the element of the state to permute during the partial rounds

val alpha : Z.t

The exponent to be used in the sbox