Module Octez_bls12_381_polynomial

type scalar = Bls12_381.Fr.t

module Internal_for_tests : sig ... end

module Domain : Domain.Domain_sig with type scalar = scalar

module Polynomial : Polynomial.Polynomial_sig with type scalar = scalar

This library implements polynomials of Bls12_381.Fr as arrays of contiguous memory in C, allowing much better performances for algorithms that scan the polynomials.

module type Evaluations_sig = Evaluations.Evaluations_sig with type scalar = scalar and type domain = Domain.t and type polynomial = Polynomial.t

module Evaluations : Evaluations.Evaluations_sig with type scalar = scalar and type domain = Domain.t and type polynomial = Polynomial.t

module type Srs_sig = Srs.S with type polynomial = Polynomial.t

module Srs : sig ... end

module G1_carray : sig ... end

module G2_carray : sig ... end