Module Octez_srs_extraction.Libsrs

module G1 = Bls12_381.G1

module G2 = Bls12_381.G2

module G1_carray = Octez_bls12_381_polynomial.G1_carray

module G2_carray = Octez_bls12_381_polynomial.G2_carray

module Checks : sig ... end

val exact_log2 : int -> int

module Srs : sig ... end

The binary format of our Srs is a sequence of G1 or G2 elements.

module Powers_of_tau : sig ... end

This module handles the file format of the result of powers-of-tau ceremony.