Module Libsrs.Powers_of_tau

This module handles the file format of the result of powers-of-tau ceremony.

The layout of the file can be found at the bottom of this file https://github.com/ebfull/powersoftau/blob/master/src/bin/verify.rs to be let tau_powers_length = 1 lsl power in G1.size_in_bytes (* alpha in g1 *)

G1.size_in_bytes (* beta in g1 *)

G2.size_in_bytes (* beta in g2 *)

(tau_powers_length * G1.size_in_bytes) (* g1_coeffs *)

(tau_powers_length * G2.size_in_bytes) (* g2_coeffs *)

(tau_powers_length * G1.size_in_bytes) (* g1_alpha_coeffs *)

(tau_powers_length * G1.size_in_bytes) (* g1_beta_coeffs *)

((tau_powers_length - 1) * G1.size_in_bytes) (* h *)

val generate_domain : int -> Bls12_381.Fr.t array

val power_of_radixfile : string -> int

val to_gs : string -> int -> int -> (bytes -> 'a) -> (domain:Octez_bls12_381_polynomial.Domain.t -> points:'b -> 'c) -> ('a array -> 'b) -> 'c

val to_g1s : string -> G1_carray.t

val to_g2s : string -> G2_carray.t