Module Plompiler.Linear_algebra

module type Ring_sig = sig ... end

module type Field_sig = sig ... end

module type Module_sig = sig ... end

This refers to the mathematical generalization of vector space called "module", where the field of scalars is replaced by a ring

module type VectorSpace_sig = sig ... end

module Make_Module (Ring : Ring_sig) : Module_sig with type t = Ring.t

module Make_VectorSpace (Field : Field_sig) : VectorSpace_sig with type t = Field.t