Module Plonk

module Arithmetic_gates : sig ... end

module Boolean_gates : sig ... end

module Circuit : sig ... end

module Cq : sig ... end

module Custom_gates : sig ... end

module Ecc_gates : sig ... end

module Evaluations_map : sig ... end

module Gates_common : sig ... end

module Hash_gates : sig ... end

module Identities : sig ... end

module Input_commitment : sig ... end

module List : sig ... end

module Main_protocol : sig ... end

aPlonK is a PlonK-based proving system. As such, it provides a way to create succinct cryptographic proofs about a given predicate, which can be then verified with a low computational cost.

module Main_protocol_intf : sig ... end

module Mod_arith_gates : sig ... end

module Permutation_gate : sig ... end

module Plookup_gate : sig ... end

module Polynomial_protocol : sig ... end

A polynomial protocol allows a prover to convince a verifier of the fact that certain algebraic identites between polynomials (polynomials that have been previously committed) hold when evaluated over a set of points. (In our implementation such set of points must be a subgroup of roots of unity.)

module Range_check_gate : sig ... end