Module Polynomial

type natural_with_infinity =

| Natural of int

| Infinity

module type UNIVARIATE = sig ... end

val generate_evaluation_domain : (module Bls12_381.Ff_sig.PRIME with type t = 'a) -> int -> 'a -> 'a array

generate_evaluation_domain (module Fp) n generator generates the domain g^{i} to be used in FFT related algorithms. generator must be a n-th principal root of unity in the finite field Fp

module MakeUnivariate (R : Bls12_381.Ff_sig.PRIME) : UNIVARIATE with type scalar = R.t

Make(Fp) builds a module of type T where the coefficients are in the prime field Fp

module Utils : sig ... end