Module Resto_json

A wrapper around json-data-encoding that exposes the modules expected by Resto.

module Encoding : Resto.ENCODING with type 'a t = 'a Json_encoding.encoding and type schema = Json_schema.schema

Encoding exposes the minimal part Json_encoding that allow to construct new encodings as well as pre-made encodings for values useful to Resto.

module type VALUE = sig ... end

A VALUE module allows the actual conversion of values between different representations. It is intended as a companion to the Encoding module above.

module Ezjsonm : VALUE with type t = Json_repr.Ezjsonm.value and type 'a encoding := 'a Encoding.t

module Bson : VALUE with type t = Json_repr_bson.bson and type 'a encoding := 'a Encoding.t