Module type Resto_json.VALUE

A VALUE module allows the actual conversion of values between different representations. It is intended as a companion to the Encoding module above.

type t

type 'a encoding

val construct : 'a encoding -> 'a -> t

val destruct : 'a encoding -> t -> 'a