Module Tezos_base

module Bitset : sig ... end

module Block_header : sig ... end

module Block_locator : sig ... end

A locator t is a data structure which roughly represents a list of block hashes in the chain. These hashes go from the top of the chain to the bottom. It is sparse in the sense that the distance between two hashes increases exponentially when we move away from the head.

module Bounded : sig ... end

This module implements bounded (or refined) versions of data types.

module Dal_srs : sig ... end

module Data_encoding_wrapper : sig ... end

module Fitness : sig ... end

module Genesis : sig ... end

module Internal_event_config : sig ... end

module Mempool : sig ... end

Tezos Shell Module - Mempool, a.k.a. the operations safe to be broadcast.

module Operation : sig ... end

module P2p_addr : sig ... end

module P2p_connection : sig ... end

module P2p_identity : sig ... end

module P2p_params : sig ... end

Types of the abstract data types (ADT) which parameterize the P2p layer.

module P2p_peer : sig ... end

module P2p_peer_id : sig ... end

module P2p_point : sig ... end

module P2p_rejection : sig ... end

module P2p_stat : sig ... end

Bandwidth usage statistics

module Point_parser : sig ... end

module Profiler : sig ... end

module Protocol : sig ... end

module S : sig ... end

module Sized : sig ... end

module Skip_list : sig ... end

This module provides an implementation of the skip list data structure.

module Test_chain_status : sig ... end

module Time : sig ... end

Time management

module TzPervasives : sig ... end

module User_activated : sig ... end

module Worker_intf : sig ... end

module Worker_types : sig ... end