Module Tezos_base.Dal_srs

val find_trusted_setup_files : ?getenv_opt:(string -> string option) -> ?getcwd:(unit -> string) -> ?file_exists:(string -> bool) -> unit -> (string * string) Tezos_error_monad.Error_monad.tzresult

find_trusted_setup_files returns the path of the two files necessary to initialize cryptographic primitives used by the DAL. See Tezos_crypto_dal.Cryptobox.