Module Tezos_base.Mempool

Tezos Shell Module - Mempool, a.k.a. the operations safe to be broadcast.

type t = {

known_valid : Tezos_crypto.Hashed.Operation_hash.Set.t;
(*
Set of valid operations on top of the current head.
*)

pending : Tezos_crypto.Hashed.Operation_hash.Set.t;
(*
Set of known not-invalid operation.
*)

}

type mempool = t

val encoding : mempool Data_encoding.t

val bounded_encoding : ?max_operations:int -> unit -> mempool Data_encoding.t

val empty : mempool

Empty mempool.

val is_empty : mempool -> bool

is_empty mempool returns true if and only if mempool is empty.

val cons_valid : Tezos_crypto.Hashed.Operation_hash.t -> mempool -> mempool

cons_valid oph t prepends oph to the known_valid field of t.

val remove : Tezos_crypto.Hashed.Operation_hash.t -> mempool -> mempool

Remove an operation from all the fields of a mempool.