Module P2p_connection.Info

Information about a connection

type 'meta t = {

incoming : bool;

peer_id : P2p_peer_id.t;

id_point : Id.t;
(*
Contact point: host address along with the TCP port on which the peer can be reached (listening port).
*)

remote_socket_port : P2p_addr.port;
(*
Port we're actually connected to. This port is equal to the contact (listening) port in case of an outgoing connection, and may be different in case of an incoming connection.
*)

announced_version : Tezos_version.Network_version.t;

private_node : bool;

local_metadata : 'meta;

remote_metadata : 'meta;

}

val pp : (Stdlib.Format.formatter -> 'meta -> unit) -> Stdlib.Format.formatter -> 'meta t -> unit

val encoding : 'meta Data_encoding.t -> 'meta t Data_encoding.t