Module P2p_connection.P2p_event

type t =

| Too_few_connections

| Too_many_connections

| Maintenance_started

| Maintenance_ended

| New_point of P2p_point.Id.t

| New_peer of P2p_peer_id.t

| Gc_points
(*
Garbage collection of known point table has been triggered.
*)

| Gc_peer_ids
(*
Garbage collection of known peer_ids table has been triggered.
*)

| Incoming_connection of P2p_point.Id.t
(*
We accept(2)-ed an incoming connection
*)

| Outgoing_connection of P2p_point.Id.t
(*
We connect(2)-ed to a remote endpoint
*)

| Authentication_failed of P2p_point.Id.t
(*
Remote point failed authentication
*)

| Accepting_request of P2p_point.Id.t * Id.t * P2p_peer_id.t
(*
We accepted a connection after authentifying the remote peer.
*)

| Rejecting_request of P2p_point.Id.t * Id.t * P2p_peer_id.t
(*
We rejected a connection after authentifying the remote peer.
*)

| Request_rejected of P2p_point.Id.t * (Id.t * P2p_peer_id.t) option
(*
The remote peer rejected our connection.
*)

| Connection_established of Id.t * P2p_peer_id.t
(*
We successfully established a authentified connection.
*)

| Bootstrap_received of {

source : P2p_peer_id.t;

}
(*
A bootstrap message has been received.
*)

| Bootstrap_sent of {

source : P2p_peer_id.t;

}
(*
A bootstrap message has been sent.
*)

| Advertise_received of {

source : P2p_peer_id.t;

}
(*
An advertise message has been received.
*)

| Advertise_sent of {

source : P2p_peer_id.t;

}
(*
An advertise message has been sent.
*)

| Swap_request_received of {

source : P2p_peer_id.t;

}
(*
A swap request has been received.
*)

| Swap_ack_received of {

source : P2p_peer_id.t;

}
(*
A swap ack has been received
*)

| Swap_request_sent of {

source : P2p_peer_id.t;

}
(*
A swap request has been sent
*)

| Swap_ack_sent of {

source : P2p_peer_id.t;

}
(*
A swap ack has been sent
*)

| Swap_request_ignored of {

source : P2p_peer_id.t;

}
(*
A swap request has been ignored
*)

| Swap_success of {

source : P2p_peer_id.t;

}
(*
A swap operation has succeeded
*)

| Swap_failure of {

source : P2p_peer_id.t;

}
(*
A swap operation has failed
*)

| Disconnection of P2p_peer_id.t
(*
We decided to close the connection.
*)

| External_disconnection of P2p_peer_id.t
(*
The connection was closed for external reason.
*)

val pp : Stdlib.Format.formatter -> t -> unit

val encoding : t Data_encoding.t