Module P2p_peer.Pool_event

type kind =

| Accepting_request
(*
We accepted a connection after authentifying the remote peer.
*)

| Rejecting_request
(*
We rejected a connection after authentifying the remote peer.
*)

| Request_rejected
(*
The remote peer rejected our connection.
*)

| Connection_established
(*
We successfully established a authentified connection.
*)

| Disconnection
(*
We decided to close the connection.
*)

| External_disconnection
(*
The connection was closed for external reason.
*)

type t = {

kind : kind;

timestamp : Time.System.t;

point : P2p_connection.Id.t;

}

val encoding : t Data_encoding.t