Module P2p_point.Pool_event

type kind =

| Outgoing_request
(*
We initiated a connection.
*)

| Accepting_request of P2p_peer_id.t
(*
We accepted a connection after authentifying the remote peer.
*)

| Rejecting_request of P2p_peer_id.t
(*
We rejected a connection after authentifying the remote peer.
*)

| Request_rejected of P2p_peer_id.t option
(*
The remote peer rejected our connection.
*)

| Connection_established of P2p_peer_id.t
(*
We successfully established a authentified connection.
*)

| Disconnection of P2p_peer_id.t
(*
We decided to close the connection.
*)

| External_disconnection of P2p_peer_id.t
(*
The connection was closed for external reason.
*)

type t = kind Time.System.stamped

val encoding : t Data_encoding.t