Module type Profiler.DRIVER

Driver

The Driver is a signature that, when instantiated, specifies the behaviour of any profiler. This means that all profilers should implement this set of functions with their specifities.

type config

Parameters to launch an instance of the driver.

type state

Internal state of an instance of the driver.

val kind : (config, state) kind

A typed kind for downcasting.

val create : config -> state

Create an instance from a config.

val time : state -> time

Gives the current time in seconds.

val record : state -> verbosity -> id -> unit

Open a sequence in the current sequence. If currently aggregating (not all aggregation scopes are closed), this has the same semantics as aggregate instead.

val aggregate : state -> verbosity -> id -> unit

Open an aggregation node in the current sequence.

val stop : state -> unit

Close the most recently opened sequence or aggregation scope.

val stamp : state -> verbosity -> id -> unit

Record a timestamp in the most recently opened sequence.

val mark : state -> verbosity -> ids -> unit

Count this event's occurences in the most recent sequence.

val span : state -> verbosity -> span -> ids -> unit

Sum the time spent in this event in the most recent sequence.

val inc : state -> report -> unit

Include a report in the current sequence.

val report : state -> report option

Consume the last toplevel report, if any.

val close : state -> unit

Flush and/or close output.