Module Tezos_base_test_helpers.Tztest

tztest contains the definition of tztest: a wrapper for tests cases running in the error monad. It also adds helpers for working with the mock sink.

val tztest : string -> Octez_alcotezt.Alcotest.speed_level -> (unit -> (unit, Tezos_base.TzPervasives.tztrace) Stdlib.result Lwt.t) -> unit Octez_alcotezt.Alcotest_lwt.test_case

Transform a function running in the error monad into an Alcotest, taking care of failing the test if the function results is an error. Note that the given function must still take care of test assertions, e.g. by using Alcotest.check.

val tztest_qcheck : ?count:int -> name:string -> 'a QCheck.arbitrary -> ('a -> ('b, Tezos_base.TzPervasives.tztrace) Stdlib.result Lwt.t) -> unit Octez_alcotezt.Alcotest_lwt.test_case

val tztest_qcheck2 : ?print:'a QCheck2.Print.t -> ?count:int -> name:string -> 'a QCheck2.Gen.t -> ('a -> ('b, Tezos_base.TzPervasives.tztrace) Stdlib.result Lwt.t) -> unit Octez_alcotezt.Alcotest_lwt.test_case

val mock_sink : Tezos_event_logging_test_helpers.Mock_sink.t Tezos_base.TzPervasives.Internal_event.sink_definition

val with_empty_mock_sink : (unit -> unit Lwt.t) -> unit Lwt.t

with_empty_mock_sink f executes f after activating or clearing a Mock_sink sink.

Sinks can only be registered and activated once, and not removed thereafter.