Parameter Make.P

type parameters

The type of parameters, which contains the necessary information to initialize the external process.

type 'response request

The type of requests understood by the external process. The external injector answers with responses of type `response for requests of type `response request.

val name : string

The name of the external process as it should appear in events. Can contain alphanumerical characters and underscores.

val request_pp : Stdlib.Format.formatter -> 'response request -> unit

A pretty printer for requests.

val internal_events : parameters -> Tezos_base.Internal_event_config.t

An accessor for the internal events configuration from the parameters.

val magic : Tezos_base.TzPervasives.Bytes.t

Magic bytes used for the handshake between the launcher and the external process.

val parameters_encoding : parameters Tezos_base.TzPervasives.Data_encoding.t

Encoding for the parameters.

type packed_request =

| Erequest : _ request -> packed_request

Existential type of requests, used for encoding.

val request_encoding : packed_request Tezos_base.TzPervasives.Data_encoding.t

Encoding for (packed) requests.

val result_encoding : 'response request -> 'response Tezos_base.TzPervasives.Data_encoding.t

result_encoding request is the encoding for results in response to the request request.

val reconfigure_event_logging_request : Tezos_base.Internal_event_config.t -> unit request

The request to reconfigure the event logging system of the external process.

val terminate_request : packed_request

The request to terminate the external process.

val socket_path_prefix : string

The prefix for the external process socket filename.

Do not use it directly except for documentation purposes; use socket_path instead.

val socket_path : socket_dir:string -> pid:int -> string

Get the external process socket path.

socket_dir is the directory where the file should be put. pid is the process ID of the external process.

val command_line_args : parameters -> socket_dir:string -> string option * string list

command_line_args params ~socket_dir is (name, args) where name is the new name for the external process (zeroth argument) (or None if it should reuse the same name as the caller) and args are the arguments that should be given to the external process executable to start.