Module type External_process_main.EXTERNAL_PROCESSING

Module type for argument of functor Make which provides the handler code for each request.

type parameters

type 'response request

type state

The type of state for the processing code of the external process.

val initial_state : parameters -> state Tezos_base.TzPervasives.tzresult Lwt.t

The initial state for the external processing.

val handle_request : parameters -> state -> 'response request -> [ `Continue of ('response * (Tezos_base.TzPervasives.Profiler.report option * Tezos_base.TzPervasives.Profiler.report option) option) Tezos_base.TzPervasives.tzresult * state | `Stop ] Lwt.t

handle_request params state acc req handles the request req and returns `Continue (res, state) if the external process should continue handling requests normally, where res is the result and state the new state. It returns `Stop if the external process should stop after processing req.