Module Tezos_base_unix.Simple_profiler

val time : unit -> Tezos_base.TzPervasives.Profiler.time

type state

val empty : Tezos_base.TzPervasives.Profiler.verbosity -> state

module Base (P : sig ... end) : sig ... end

The Base functor helps to define other backends without having to write the same functions again and again.

val headless : Tezos_base.TzPervasives.Profiler.verbosity Tezos_base.TzPervasives.Profiler.driver

Memory driver allowing building intermediate reports with purpose of being included in external reports or manipulated directly.

val auto_write_as_txt_to_file : (string * Tezos_base.TzPervasives.Profiler.verbosity) Tezos_base.TzPervasives.Profiler.driver

Driver printing its report in plain text to a file whenever a toplevel section ends.

val auto_write_as_json_to_file : (string * Tezos_base.TzPervasives.Profiler.verbosity) Tezos_base.TzPervasives.Profiler.driver

Driver printing its report in JSON to a file whenever a toplevel section ends.