Module Simple_profiler.Base

The Base functor helps to define other backends without having to write the same functions again and again.

Given a way to get and set a state and an output function, the functor will produce a module implementing the DRIVER interface.

Note that the produced module will try to output report every time stamp, inc, mark, span or stop is used.

Parameters

module P : sig ... end

Signature

val time : P.t -> Tezos_base.TzPervasives.Profiler.time

val record : P.t -> Tezos_base.TzPervasives.Profiler.verbosity -> Tezos_base.TzPervasives.Profiler.id -> unit

val aggregate : P.t -> Tezos_base.TzPervasives.Profiler.verbosity -> Tezos_base.TzPervasives.Profiler.id -> unit

val report : P.t -> Tezos_base.TzPervasives.Profiler.report option

val stamp : P.t -> Tezos_base.TzPervasives.Profiler.verbosity -> Tezos_base.TzPervasives.Profiler.id -> unit

val inc : P.t -> Tezos_base.TzPervasives.Profiler.report -> unit

val mark : P.t -> Tezos_base.TzPervasives.Profiler.verbosity -> Tezos_base.TzPervasives.Profiler.ids -> unit

val span : P.t -> Tezos_base.TzPervasives.Profiler.verbosity -> Tezos_base.TzPervasives.Profiler.span -> Tezos_base.TzPervasives.Profiler.ids -> unit

val stop : P.t -> unit