Module Tezos_crypto.Helpers

module MakeRaw (H : sig ... end) : sig ... end

module MakeB58 (H : sig ... end) : sig ... end

module MakeEncoder (H : sig ... end) : sig ... end

module MakeIterator (H : sig ... end) : sig ... end

module Make (H : sig ... end) : sig ... end

val yes_crypto_environment_variable : string

This variable is used to enable yes-cryptography at runtime.

With yes-cryptography activated, signatures are faked and always considered valid.

This can put your software at risk of being

considered faulty/malicious if it fake signs,

exploited by attackers if it fake-checks signatures.

This should be used for testing purposes only and/or with extreme care.

type yes_crypto_kind =

| Fast

| Yes

| No

val yes_crypto_kind : yes_crypto_kind

val is_yes_crypto_enabled : bool

module Events : sig ... end