Module Signature.Of_V_latest

Converting from signatures of V_latest to V_latest. This module implements conversions which are the identity, so total, but we keep the signature as CONV_OPT for compatibility with V0.Of_V_latest and V1.Of_V_latest and to ease snapshotting.

Converting from signatures of V_latest to V_latest. This module implements conversions which are the identity, so total, but we keep the signature as CONV_OPT for compatibility with V0.Of_V_latest and V1.Of_V_latest and to ease snapshotting.

val public_key_hash : Signature.V_latest.Public_key_hash.t -> Signature.V_latest.Public_key_hash.t option

val public_key : Signature.V_latest.Public_key.t -> Signature.V_latest.Public_key.t option

val secret_key : Signature.V_latest.Secret_key.t -> Signature.V_latest.Secret_key.t option

val signature : Signature.V_latest.t -> Signature.V_latest.t option