Module type Cryptobox_intf.VERIFIER

type t

A precomputed set of constants

type parameters = Tezos_crypto_dal_octez_dal_config.Dal_config.parameters = {

redundancy_factor : int;

page_size : int;

slot_size : int;

number_of_shards : int;

}

Parameters to build a value of type t

type ('a, 'b) error_container = {

given : 'a;

expected : 'b;

}

val parameters_encoding : parameters Data_encoding.t

An encoding for values of type parameters.

val make : parameters -> (t, [> `Fail of string ]) Stdlib.result

make precomputes the set of values needed by the cryptographic primitives defined in this module and stores them in a value of type t

val parameters : t -> parameters

parameters t returns the parameters given when t was initialised with the function make

type commitment

Commitment to a polynomial.

module Commitment : COMMITMENT with type t := commitment

type commitment_proof

A proof that the polynomial associated to some commitment is bounded by a constant.

module Commitment_proof : Kzg.Interfaces.Degree_check_proof with type t := commitment_proof

val verify_commitment : t -> commitment -> commitment_proof -> bool

verify_commitment t commitment proof returns true if and only if the size of the data committed via commitment does not exceed the slot_size declared in t.

The verification time is constant.

type page = bytes

The original slot can be split into a list of pages of fixed size. This size is given by the parameter page_size given to the function make.

type page_proof

A proof that the evaluation of points of a polynomial is part of a commitment.

val page_proof_encoding : page_proof Data_encoding.t

An encoding for the proof of a page.

val pages_per_slot : parameters -> int

pages_per_slot t returns the number of expected pages per slot.

val verify_page : t -> commitment -> page_index:int -> page -> page_proof -> (unit, [> `Invalid_degree_strictly_less_than_expected of (int, int) error_container | `Invalid_page | `Page_length_mismatch | `Page_index_out_of_range ]) Stdlib.Result.t

verify_page t commitment ~page_index page proof returns Ok () if the proof certifies that the page is the page_index-th page of the slot with the given commitment.

Fails with:

Error `Invalid_page if the verification Fails

Error `Invalid_degree_strictly_less_than_expected _ if the SRS contained in t is too small to proceed with the verification

Error `Page_length_mismatch if the page is not of the expected length page_size given for the initialisation of t

Error `Page_index_out_of_range if page_index is out of the range 0, slot_size/page_size - 1 where slot_size and page_size are given for the initialisation of t

Ensures:

verify_page t commitment ~page_index page proof = Ok () if and only if page = Bytes.sub slot (page_index * t.page_size) t.page_size), proof = prove_page t polynomial page_index, p = polynomial_from_slot t slot, and commitment = commit t p.

type share

val share_encoding : share Data_encoding.t

type shard = {

index : int;

share : share;

}

val shard_encoding : shard Data_encoding.t

type shard_proof

val shard_proof_encoding : shard_proof Data_encoding.t

val verify_shard : t -> commitment -> shard -> shard_proof -> (unit, [> `Invalid_degree_strictly_less_than_expected of (int, int) error_container | `Invalid_shard | `Shard_length_mismatch | `Shard_index_out_of_range of string ]) Stdlib.Result.t