Module Tezos_crypto_dal_octez_dal_config.Dal_config

node parameters for the DAL.

type parameters = {

redundancy_factor : int;

page_size : int;

slot_size : int;

number_of_shards : int;

}

val parameters_encoding : parameters Data_encoding.t

type t = {

activated : bool;
(*
true if the DAL is activated. This may have an impact on the loading time of the node.
*)

bootstrap_peers : string list;
(*
List of bootstrap peers to use in the DAL p2p.
*)

}

val encoding : t Data_encoding.t

val default : t

The default configuration is {activated = true; bootstrap_peers = []}.