Module Tezos_error_monad.Monad_maker

module type S = sig ... end

S is the Tezos-specific extension to the generic monad provided by Lwtreslib. It sets some defaults (e.g., it defaults traced failures), it brings some qualified identifiers into the main unqualified part (e.g., return_unit), it provides some tracing helpers and some in-monad assertion checks.

module Make (Error : sig ... end) (Trace : Sig.TRACE) (Monad : Tezos_lwt_result_stdlib.Lwtreslib.TRACED_MONAD with type 'error trace := 'error Trace.trace) : S with type error := Error.error and type 'error trace := 'error Trace.trace