Module Tezos_error_monad.Sig

module type ERROR_CATEGORY = sig ... end

module type PREFIX = sig ... end

module type CORE = sig ... end

module type WITH_WRAPPED = sig ... end

module type TRACE = sig ... end

module type MONAD_EXTENSION = sig ... end

MONAD_EXTENSION is the Tezos-specific extension to the generic monad provided by Lwtreslib. It sets some defaults (e.g., it defaults traced failures), it brings some qualified identifiers into the main unqualified part (e.g., return_unit), it provides some tracing helpers and some in-monad assertion checks.