Module Internal_event.Lwt_worker_logger

The worker logger is meant for use with Lwt_utils.worker.

val on_event : string -> [ `Started | `Ended | `Failed of string ] -> unit Lwt.t

on_event status emits an event of type t and matches the signature required by Lwt_utils.worker.