Module Subconfig.Message

include Tezos_base.TzPervasives.PRINTABLE

type t

val pp : Stdlib.Format.formatter -> t -> unit

val valid : ?message:t -> message_id:Message_id.t -> unit -> [ `Valid | `Unknown | `Outdated | `Invalid ]

valid performs an application layer-level validity check on a message id and a message if given.

The message id (and message if any) could either be `Valid in the current context or `Invalid, meaning that it is/they are not valid (in the present time, in the past and in the future). The application layer could also return `Outdated or `Unknown if the message id is outdated or if the application doesn't care about validity. In this case, the application might omit some costly validity checks.