Module Worker.Introspection

Introspection and stats facilities

type stats = private {

mutable count_topics : int64;
(*
Counts the number of topics of the node. It's the diff between Join and Leave topics events.
*)

mutable count_connections : int64;
(*
Counts the number of connections of the node. It's the diff between New_connection and Disconnection events.
*)

mutable count_bootstrap_connections : int64;
(*
Counts the number of connections of the node to bootstrap peers. It's a refinement of count_connections for when the remote peer declares itself as a bootstrap peer.
*)

mutable count_sent_app_messages : int64;
(*
Count sent app messages.
*)

mutable count_sent_grafts : int64;
(*
Count sent grafts.
*)

mutable count_sent_prunes : int64;
(*
Count sent prunes.
*)

mutable count_sent_ihaves : int64;
(*
Count sent ihaves.
*)

mutable count_sent_iwants : int64;
(*
Count sent iwants.
*)

mutable count_recv_valid_app_messages : int64;
(*
Count received app messages that are known to be valid.
*)

mutable count_recv_invalid_app_messages : int64;
(*
Count received app messages that are known to be invalid.
*)

mutable count_recv_unknown_validity_app_messages : int64;
(*
Count received app messages we won't validate.
*)

mutable count_recv_outdated_app_messages : int64;
(*
Count received app messages that are outdated.
*)

mutable count_recv_grafts : int64;
(*
Count successfully received & processed grafts.
*)

mutable count_recv_prunes : int64;
(*
Count successfully received & processed prunes.
*)

mutable count_recv_ihaves : int64;
(*
Count successfully received & processed ihaves.
*)

mutable count_recv_iwants : int64;
(*
Count successfully received & processed iwants.
*)

}

A record containing some stats about what happened in the Gossipsub worker.

val empty_stats : unit -> stats