Module Immutable_chunked_byte_vector.Chunk

type t

Chunk within the byte vector

val of_bytes : bytes -> t

Create a chunk and copy the given bytes into it.

val to_bytes : t -> bytes

Copy the contents of a chunk into a fresh bytes.

val size : int64

Size of a chunk in bytes - with 9 bits of address space the chunk is 512 bytes

val num_needed : int64 -> int64

num_needed len Computes the number of chunks needed to cover len.

val encoding : t Tezos_tree_encoding.t

encoding is a Tezos_tree_encoding for Chunk.