Module type Lazy_map.KeyS

KeyS is the qualifier signature for key types in the lazy map. Externally visible and accessible keys of the lazy map are always non-negative. However, the lazy map implementation may internally use negative keys therefore modules of type KeyS must support them.

include Stdlib.Map.OrderedType

type t

val compare : t -> t -> int

val to_string : t -> string