Module Lazy_vector.Mutable

Make generates a lazy vector module using a given Key module.

module type ImmutableS = S

module type S = sig ... end

S is the signature of a mutable lazy vector module.

module Make (Vector : ImmutableS) : S with type key = Vector.key and module Vector = Vector

module IntVector : S with type key = int and module Vector = IntVector

module Int32Vector : S with type key = int32 and module Vector = Int32Vector

module Int64Vector : S with type key = int64 and module Vector = Int64Vector

module ZVector : S with type key = Z.t and module Vector = ZVector