Module type Lazy_vector.KeyS

KeyS is the qualifier signature for key types in the lazy vector. Externally visible and accessible keys of the lazy vector are always non-negative. However, the lazy vector implementation may internally use negative keys therefore modules of type KeyS must support them.

include Stdlib.Map.OrderedType

type t

val compare : t -> t -> int

val unsigned_compare : t -> t -> int

val zero : t

val add : t -> t -> t

val sub : t -> t -> t

val pred : t -> t

val succ : t -> t

val to_string : t -> string