Module Tezos_micheline.Micheline_encoding

val table_encoding : variant:string -> 'l Data_encoding.encoding -> 'p Data_encoding.encoding -> ('l, 'p) Micheline.node Data_encoding.encoding

Encoding for expressions, as their canonical encoding. Locations are stored in a side table. See canonical_encoding for the variant parameter.

val erased_encoding : variant:string -> 'l -> 'p Data_encoding.encoding -> ('l, 'p) Micheline.node Data_encoding.encoding

Encoding for expressions, as their canonical encoding. Locations are erased when serialized, and restored to a provided default value when deserialized. See canonical_encoding for the variant parameter.

val node_encoding : Micheline_parser.node Data_encoding.encoding

val canonical_location_encoding : Micheline.canonical_location Data_encoding.encoding

Encoding for canonical integer locations.

val canonical_encoding : variant:string -> 'l Data_encoding.encoding -> 'l Micheline.canonical Data_encoding.encoding

Encoding for expressions in canonical form. The first parameter is a name used to produce named definitions in the schemas. Make sure to use different names if two expression variants with different primitive encodings are used in the same schema.

val canonical_encoding_v0 : variant:string -> 'l Data_encoding.encoding -> 'l Micheline.canonical Data_encoding.encoding

Old version of canonical_encoding for backward compatibility. Do not use in new code.

val canonical_encoding_v1 : variant:string -> 'l Data_encoding.encoding -> 'l Micheline.canonical Data_encoding.encoding

Old version of canonical_encoding for backward compatibility. Do not use in new code.

val canonical_encoding_v2 : variant:string -> 'l Data_encoding.encoding -> 'l Micheline.canonical Data_encoding.encoding

Alias for canonical_encoding.