Module Tezos_p2p_services

module Connection_metadata : sig ... end

All the metadata associated to a running connection.

module P2p_errors : sig ... end

module P2p_limits : sig ... end

The configurable constants used by the p2p layer as maximum, with their encoding and default values.

module P2p_services : sig ... end

module Peer_metadata : sig ... end

All the (persistent) metadata associated to a peer.

module Point_reconnection_config : sig ... end

Parameters of the heuristic determining waiting time between reconnection to a neighbour that misbehaved