Module P2p_services.ACL

type ip_list = {

ips : Ipaddr.V6.t list;
(*
If not_reliable_since is None contains the list of currently greylisted IPs.
*)

not_reliable_since : Ptime.t option;
(*
Contains the date of the first time that the list of IPs has been overflowed and became not reliable.
*)

}

Structure that provides a friendly and bounded list of currently greylisted IPs.

val clear : Tezos_rpc.Context.simple -> unit -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val get_greylisted_peers : Tezos_rpc.Context.simple -> Tezos_base.TzPervasives.P2p_peer.Id.t list Tezos_base.TzPervasives.tzresult Lwt.t

val get_greylisted_ips : Tezos_rpc.Context.simple -> ip_list Tezos_base.TzPervasives.tzresult Lwt.t

module S : sig ... end