Module Tezos_profiler_backend.Opentelemetry_profiler

val initialize : ?unique_identifier:string -> string -> unit

initialize ?unique_identifier service_name initializes the Opentelemetry client to output the traces. ?unique_identifier can be used to discriminate a service used in multiple binaries (for example multiple `octez-node` in the network).

val op_hash_to_trace_id : Tezos_crypto.Hashed.Operation_hash.t -> Opentelemetry.Trace_id.t

op_hash_to_trace_id op_hash returns a valid Opentelemetry.Trace_id.t that can be used when tracing. When tracing an operation application, using the operation hash as trace id can serve to produce a consistent trace amongst multiple nodes to view the operation propagation.

val add_event : ?attrs:Opentelemetry.key_value list -> string -> unit

add_event ?attrs name adds an event to the currently opened scope. If no scope has been opened yet (i.e. not called in the continuation of Opentelemetry.Trace.with_), the function does nothing.

val trace : ?force_new_trace_id:bool -> ?trace_state:string -> ?service_name:string -> ?attrs:(string * Opentelemetry.value) Tezos_base.TzPervasives.trace -> ?kind:Opentelemetry_proto.Trace.span_span_kind -> ?trace_id:Opentelemetry.Trace_id.t -> ?parent:Opentelemetry.Span_id.t -> ?scope:Opentelemetry.Scope.t -> ?links:Opentelemetry_proto.Trace.span_link list -> string -> (unit -> 'a) -> 'a

trace id k is Opentelemetry.Trace.with_ id (fun _ -> k ()). It doesn't give the scope to the callback, as it is meant to be called by custom functions.

val trace_operation : [< `Hash of Tezos_base.TzPervasives.Operation_hash.t | `Operation of Tezos_base.Operation.t ] -> ?attrs:(string * Opentelemetry.value) Tezos_base.TzPervasives.trace -> ?force_new_trace_id:bool -> ?trace_state:string -> ?service_name:string -> ?kind:Opentelemetry_proto.Trace.span_span_kind -> ?parent:Opentelemetry.Span_id.t -> ?scope:Opentelemetry.Scope.t -> ?links:Opentelemetry_proto.Trace.span_link list -> string -> (unit -> 'a) -> 'a

trace_operation op id k uses the operation hash as a trace_id and calls trace ~trace_id id k.

val update_scope : Opentelemetry.Scope.t option -> (unit -> 'a) -> 'a

update_scope s f updates the Opentelemetry scope in the ambient context before calling f.

type config = {

service_name : string;

verbosity : Tezos_base.Profiler.verbosity;

}

module Driver : Tezos_base.Profiler.DRIVER with type config = config

Mocked driver, that serves only to register the opentelemetry as a valid profiler and enable relevant functions with the ppx.

val opentelemetry : config Tezos_base.Profiler.driver