Module Opentelemetry_profiler.Driver

Mocked driver, that serves only to register the opentelemetry as a valid profiler and enable relevant functions with the ppx.

type config = config

Parameters to launch an instance of the driver.

type state

Internal state of an instance of the driver.

val kind : (config, state) Tezos_base.Profiler.kind

A typed kind for downcasting.

val create : config -> state

Create an instance from a config.

val time : state -> Tezos_base.Profiler.time

Gives the current time in seconds.

val record : state -> Tezos_base.Profiler.verbosity -> Tezos_base.Profiler.id -> unit

Open a sequence in the current sequence. If currently aggregating (not all aggregation scopes are closed), this has the same semantics as aggregate instead.

val aggregate : state -> Tezos_base.Profiler.verbosity -> Tezos_base.Profiler.id -> unit

Open an aggregation node in the current sequence.

val stop : state -> unit

Close the most recently opened sequence or aggregation scope.

val stamp : state -> Tezos_base.Profiler.verbosity -> Tezos_base.Profiler.id -> unit

Record a timestamp in the most recently opened sequence.

val mark : state -> Tezos_base.Profiler.verbosity -> Tezos_base.Profiler.ids -> unit

Count this event's occurences in the most recent sequence.

val span : state -> Tezos_base.Profiler.verbosity -> Tezos_base.Profiler.span -> Tezos_base.Profiler.ids -> unit

Sum the time spent in this event in the most recent sequence.

val inc : state -> Tezos_base.Profiler.report -> unit

Include a report in the current sequence.

val report : state -> Tezos_base.Profiler.report option

Consume the last toplevel report, if any.

val close : state -> unit

Flush and/or close output.