Module Tezos_rpc.Service

type meth = [

| `GET

| `POST

| `DELETE

| `PUT

| `PATCH

]

val string_of_meth : [< meth ] -> string

val meth_of_string : string -> [> meth ] option

val meth_encoding : meth Data_encoding.t

module MethMap = Resto.MethMap

type (+'m, 'pr, 'p, 'q, 'i, 'o, 'e) raw = ('m, 'pr, 'p, 'q, 'i, 'o, 'e) Resto.MakeService(Tezos_rpc__.RPC_encoding).t constraint 'meth = [< meth ]

type error = Tezos_error_monad.Error_monad.error list

type (+'meth, 'prefix, 'params, 'query, 'input, 'output) t = ('meth, 'prefix, 'params, 'query, 'input, 'output, error) raw constraint 'meth = [< meth ]

type (+'meth, 'prefix, 'params, 'query, 'input, 'output) service = ('meth, 'prefix, 'params, 'query, 'input, 'output, error) raw constraint 'meth = [< meth ]

include module type of struct include Resto.MakeService(Tezos_rpc__.RPC_encoding) end with type (+'m, 'pr, 'p, 'q, 'i, 'o, 'e) t := ('m, 'pr, 'p, 'q, 'i, 'o, 'e) raw and type (+'m, 'pr, 'p, 'q, 'i, 'o, 'e) service := ('m, 'pr, 'p, 'q, 'i, 'o, 'e) raw

Services.

val meth : ('meth, 'prefix, 'params, 'query, 'input, 'output, 'error) raw -> 'meth

val query : ('meth, 'prefix, 'params, 'query, 'input, 'output, 'error) raw -> 'query Resto.Query.t

val path : ('meth, 'prefix, 'params, 'query, 'input, 'output, 'error) raw -> ('prefix, 'params) Resto.Path.t

type _ input = _ Resto.MakeService(Tezos_rpc__.RPC_encoding).input =

| No_input : unit input

| Input : 'input Data_encoding.t -> 'input input

val input_encoding : ('meth, 'prefix, 'params, 'query, 'input, 'output, 'error) raw -> 'input input

val output_encoding : ('meth, 'prefix, 'params, 'query, 'input, 'output, 'error) raw -> 'output Data_encoding.t

post_service, delete_service, put_service, and patch_service are similar to get_service but for the other HTTP methods.

Note that some of these functions take an additional input argument. This is only for the services with methods that expect additional parameters. It is used internally to encode/decode additional parameters passed in a dedicated payload rather than in the path/query parameters.

val prefix : ('prefix, 'inner_prefix) Resto.Path.t -> ('meth, 'inner_prefix, 'params, 'query, 'input, 'output, 'error) raw -> ('meth, 'prefix, 'params, 'query, 'input, 'output, 'error) raw

val subst0 : ([< Resto.meth ] as 'm, 'p, 'p, 'q, 'i, 'o, 'e) raw -> ('m, 'p2, 'p2, 'q, 'i, 'o, 'e) raw

val subst1 : ([< Resto.meth ] as 'm, 'p, 'p * 'a, 'q, 'i, 'o, 'e) raw -> ('m, 'p2, 'p2 * 'a, 'q, 'i, 'o, 'e) raw

val subst2 : ([< Resto.meth ] as 'm, 'p, ('p * 'a) * 'b, 'q, 'i, 'o, 'e) raw -> ('m, 'p2, ('p2 * 'a) * 'b, 'q, 'i, 'o, 'e) raw

val subst3 : ([< Resto.meth ] as 'm, 'p, (('p * 'a) * 'b) * 'c, 'q, 'i, 'o, 'e) raw -> ('m, 'p2, (('p2 * 'a) * 'b) * 'c, 'q, 'i, 'o, 'e) raw

type ('prefix, 'params, 'error) description_service = ([ `GET ], 'prefix, 'params * string list, Resto.Description.request, unit, (Data_encoding.json_schema * Data_encoding.Binary_schema.t) Resto.Description.directory, 'error) raw

type 'input request = 'input Resto.MakeService(Tezos_rpc__.RPC_encoding).request = {

meth : Resto.meth;

uri : Uri.t;

input : 'input input;

}

val forge_request : ('meth, unit, 'params, 'query, 'input, 'output, 'error) raw -> ?base:Uri.t -> 'params -> 'query -> 'input request

val forge_partial_request : ('meth, 'prefix, 'params, 'query, 'input, 'output, 'error) raw -> ?base:Uri.t -> 'params -> 'query -> 'input request

module Internal = Resto.MakeService(Tezos_rpc__.RPC_encoding).Internal

val get_service : ?description:string -> query:'query Resto.Query.t -> output:'output Data_encoding.t -> ('prefix, 'params) Resto.Path.t -> ([ `GET ], 'prefix, 'params, 'query, unit, 'output) service

val post_service : ?description:string -> query:'query Resto.Query.t -> input:'input Data_encoding.t -> output:'output Data_encoding.t -> ('prefix, 'params) Resto.Path.t -> ([ `POST ], 'prefix, 'params, 'query, 'input, 'output) service

val delete_service : ?description:string -> query:'query Resto.Query.t -> output:'output Data_encoding.t -> ('prefix, 'params) Resto.Path.t -> ([ `DELETE ], 'prefix, 'params, 'query, unit, 'output) service

val patch_service : ?description:string -> query:'query Resto.Query.t -> input:'input Data_encoding.t -> output:'output Data_encoding.t -> ('prefix, 'params) Resto.Path.t -> ([ `PATCH ], 'prefix, 'params, 'query, 'input, 'output) service

val put_service : ?description:string -> query:'query Resto.Query.t -> input:'input Data_encoding.t -> output:'output Data_encoding.t -> ('prefix, 'params) Resto.Path.t -> ([ `PUT ], 'prefix, 'params, 'query, 'input, 'output) service