Module Wallet.Spending_key

include Core_sig.T_encoding with type t = Client.Spending_key.t

type t = Client.Spending_key.t

val encoding : t Data_encoding.t

type ovk

Outgoing viewing keys allow to keep track of spendings from a key. This feature is optional, a payer can decide to not keep track of its outgoing transactions. *

val of_bytes : bytes -> t option

val to_bytes : t -> bytes

val of_seed : bytes -> t

Generate a key from a 32 byte uniformly random seed. The source of randomness should be of cryptographic quality. *

val derive_key : t -> int32 -> t

derive_key k n derives the nth key from k *

val child_index : t -> int32

Returns the next available index from a key *

val b58check_encoding : t Tezos_crypto.Base58.encoding