Module Tezos_scoru_wasm.Data_encoding_utils

type 'a case_incr

case_incr is like Data_encoding.case but shall be used in unions where the case tags are monotonically increasing.

val case_incr : string -> 'a Tezos_base.TzPervasives.Data_encoding.t -> ('b -> 'a option) -> ('a -> 'b) -> 'b case_incr

case_incr title encoding probe extract creates a case_incr similar to Data_encoding.case ~title auto_tag encoding probe extract where the auto_tag is automatically determined later.

val unit_case_incr : string -> 'a -> 'a case_incr

unit_case_incr title value is a convenience function around case_incr to simplify construction of cases with unit-isomorphic bodies.

val union_incr : 'a case_incr list -> 'a Tezos_base.TzPervasives.Data_encoding.t

union_incr cases creates a data encoding like Data_encoding.union for case_incr items.

module Little_endian : sig ... end