Module Tezos_scoru_wasm.Host_funcs

val lookup : version:Wasm_pvm_state.version -> Tezos_webassembly_interpreter.Ast.name -> Tezos_webassembly_interpreter.Instance.extern

lookup ~version name retrieves or instantiates a host function by the given name.

Used to plug host function wrappers in the WASN interpreter linker.

val lookup_opt : version:Wasm_pvm_state.version -> Tezos_webassembly_interpreter.Ast.name -> Tezos_webassembly_interpreter.Instance.extern option

lookup_opt ~version name is exactly lookup name but returns an option instead of raising `Not_found`.

val registry : version:Wasm_pvm_state.version -> write_debug:Builtins.write_debug -> Tezos_webassembly_interpreter.Host_funcs.registry

registry ~version ~write_debug returns the host functions registry for the expected PVM version, and with the expected implementation for the write_debug function.

exception Bad_input

exception Key_too_large of int

A durable key was given by the kernel with a longer-than-allowed length.

module Error : sig ... end

module type Memory_access = sig ... end

module Memory_access_interpreter : Memory_access with type t := Tezos_webassembly_interpreter.Instance.memory_inst

module Aux : sig ... end

module Tick_model : sig ... end

Defines the tick consumption of memory access for the host functions.

module Internal_for_tests : sig ... end