Module Host_funcs.Error

type t =

| Store_key_too_large
(*
The store key submitted as an argument of a host function exceeds the authorized limit. Has code `-1`.
*)

| Store_invalid_key
(*
The store key submitted as an argument of a host function cannot be parsed. Has code `-2`.
*)

| Store_not_a_value
(*
The contents (if any) of the store under the key submitted as an argument of a host function is not a value. Has code `-3`.
*)

| Store_invalid_access
(*
An access in a value of the durable storage has failed, supposedly out of bounds of a value. Has code `-4`.
*)

| Store_value_size_exceeded
(*
Writing a value has exceeded 2^31 bytes. Has code `-5`.
*)

| Memory_invalid_access
(*
An address is out of bound of the memory. Has code `-6`.
*)

| Input_output_too_large
(*
The input or output submitted as an argument of a host function exceeds the authorized limit. Has code `-7`.
*)

| Generic_invalid_access
(*
Generic error code for unexpected errors. Has code `-8`.
*)

| Store_readonly_value
(*
A value cannot be modified if it is readonly. Has code `-9`.
*)

| Store_not_a_node
(*
There is no tree at key. It has no value, nor any subtrees. Has code `-10`.
*)

| Full_outbox
(*
The outbox is full an cannot accept new messages at this level. Has code `-11`.
*)

| Store_invalid_subkey_index
(*
Trying to get the nth subkey, where n is too big. Has code `-12`.
*)

| Store_value_already_exists
(*
Trying to create a value at a key that already has an associated value. Has code `-13`.
*)

val code : t -> int32

code error returns the error code associated to the error.