Module Host_funcs.Internal_for_tests

val metadata_size : int

The number of bytes used to store the metadata of a rollup in memory

val write_output : Tezos_webassembly_interpreter.Instance.func_inst

val read_input : Tezos_webassembly_interpreter.Instance.func_inst

val store_exists : Tezos_webassembly_interpreter.Instance.func_inst

val store_has : Tezos_webassembly_interpreter.Instance.func_inst

store_has returns whether a key corresponds to a value and/or subtrees. Namely, it returns the following enum:

0: There is no value at key, nor subtrees under key.

1: There is a value at key, but no subtrees under key.

2: There is no value at key, but there are subtrees under key.

3: There is a value at key, and subtrees under key.

val store_delete : Tezos_webassembly_interpreter.Instance.func_inst

val store_delete_value : Tezos_webassembly_interpreter.Instance.func_inst

val store_copy : Tezos_webassembly_interpreter.Instance.func_inst

val store_move : Tezos_webassembly_interpreter.Instance.func_inst

val store_value_size : Tezos_webassembly_interpreter.Instance.func_inst

val store_create : Tezos_webassembly_interpreter.Instance.func_inst

val store_read : Tezos_webassembly_interpreter.Instance.func_inst

val store_write : Tezos_webassembly_interpreter.Instance.func_inst

val store_list_size : Tezos_webassembly_interpreter.Instance.func_inst

val store_get_nth_key : Tezos_webassembly_interpreter.Instance.func_inst

val store_get_hash : Tezos_webassembly_interpreter.Instance.func_inst

val write_debug : Tezos_webassembly_interpreter.Instance.func_inst

val ec_pairing_check_bls12_381 : Tezos_webassembly_interpreter.Instance.func_inst