Module Tezos_scoru_wasm.Pvm_input_kind

type protocol =

| Nairobi

| Oxford

| ParisB

| ParisC

| Proto_alpha

| Quebec

type internal_message_kind =

| Transfer

| Start_of_level
(*
Internal message put at the beginning of each inbox's level.
*)

| End_of_level
(*
Internal message put at the end of each inbox's level.
*)

| Info_per_level
(*
Internal message containing the timestamp of the current block and the hash of the previous block.
*)

| Protocol_migration of protocol

internal_message_kind represent an internal message in a inbox.

type t =

| Internal of internal_message_kind

| External

| Other

A type representing messages from Layer 1 to Layer 2. Internal ones are originated from Layer 1 smart-contracts and external ones are messages from an external manager operation. Other messages represents non decodable tags. It aims to be future-proof and prevent discarding any new type of message..

val from_raw_input : string -> t

from_raw_input input takes a message produced by the L1 protocol and returns its kind.

module Internal_for_tests : sig ... end