Module Pvm_input_kind.Internal_for_tests

val to_binary_input : t -> string option -> string

to_binary_input kind input returns the serialized representation of an input according to its kind. Internal message payloads and external message payloads are prefixed by their tag, and `Other` messages result in an exception. These messages are meant to be used in tests only, and does not give any guarantee on their validity according to the protocol's representation.

raises Failure
on `Other` messages, and mismatches between kind and presence or absence of the input.