Module Tezos_scoru_wasm.Wasm_pvm_state

type version =

| V0

| V1

| V2

| V3

| V4

| V5

| V6

val versions : (string * version) list

val versions_flip : (version * string) list

val version_encoding : version Tezos_base.TzPervasives.Data_encoding.encoding

type input_info = {

inbox_level : Tezos_base.Bounded.Non_negative_int32.t;
(*
The inbox level at which the message exists.
*)

message_counter : Z.t;
(*
The index of the message in the inbox.
*)

}

Represents the location of an input message.

type output_info = {

outbox_level : Tezos_base.Bounded.Non_negative_int32.t;
(*
The outbox level at which the message exists.
*)

message_index : Z.t;
(*
The index of the message in the outbox.
*)

}

Represents the location of an output message.

type reveal = Tezos_webassembly_interpreter.Host_funcs.reveal =

| Reveal_raw of string

val reveal_raw_data_tag : char

val reveal_raw_data : string -> reveal

val reveal_metadata_payload : string

val reveal_metadata : reveal

module Compatibility : sig ... end

type input_request =

| No_input_required
(*
The VM does not expect any input.
*)

| Input_required
(*
The VM needs input in order to progress.
*)

| Reveal_required of reveal

Represents the state of input requests.

type info = {

current_tick : Z.t;
(*
The number of ticks processed by the VM, zero for the initial state. current_tick must be incremented for each call to step
*)

last_input_read : input_info option;
(*
The last message to be read by the VM, if any.
*)

input_request : input_request;
(*
The current VM input request.
*)

}

Represents the state of the VM.

module Internal_state : sig ... end

This module type defines the state for the PVM. For use in lib_scoru_wasm only.