Module Tezos_stdlib

module Bits : sig ... end

module Bloomer : sig ... end

module Bounded_heap : sig ... end

Bounded sequence: keep only the n greatest elements.

module Bounded_min_heap : sig ... end

Bounded min-heap: keep only the n smallest elements.

module Circular_buffer : sig ... end

This module implements a bufferisation abstraction to store temporary raw data chunks (as bytes) when chunks are read sequentially. The function write allows to store chunks in the buffer and the function read to read them from the buffer.

module Compare : sig ... end

module FallbackArray : sig ... end

This module implements arrays equipped with accessors that cannot raise exceptions. Reading out of the bounds of the arrays return a fallback value fixed at array construction time, writing out of the bounds of the arrays is ignored.

module FunctionalArray : sig ... end

This module implements functional arrays equipped with accessors that cannot raise exceptions following the same design principles as FallbackArray:

module Hash_queue : sig ... end

Bounded queues combined with hash-tables, based on Aches.

module Hex : sig ... end

module Lwt_dropbox : sig ... end

A 'dropbox' with a single element.

module Lwt_idle_waiter : sig ... end

module Lwt_pipe : sig ... end

module Lwt_utils : sig ... end

module Memory : sig ... end

module Pretty_printing : sig ... end

module Tag : sig ... end

Tags and tag sets. Tags are basically similar to a plain extensible variant type, but wrapped with metadata that enables them to be printed generically and combined into tag sets where each tag is either not present or associated with a specific value.

module TzBytes : sig ... end

module TzEndian : sig ... end

module TzFilename : sig ... end

module TzList : sig ... end

module TzString : sig ... end

module Utils : sig ... end