Module Tezos_stdlib.Bounded_min_heap

Bounded min-heap: keep only the n smallest elements.

module Make (Id : Stdlib.Hashtbl.HashedType) (E : sig ... end) : sig ... end

Make (ID) (E) creates a min heap for element of type E.t using the E.compare function. The heap is bounded and adding more elements than its capacity will fail. The module ID is used to check if an element already exists at insertion.