Module Tezos_stdlib.Circular_buffer

This module implements a bufferisation abstraction to store temporary raw data chunks (as bytes) when chunks are read sequentially. The function write allows to store chunks in the buffer and the function read to read them from the buffer.

The global contract is that if we write consecutively d1;d2 onto the buffer, then we have to fully read d1 and d2, in that order.

This contract is not enforced by the library, it is the user responsibility to respect it.

If the circular buffer is full, a new temporary buffer is allocated to store the chunk of data to be written.

type t

Type of circular buffers

type data

An abstraction over a chunk of data written in the buffer.

val create : ?maxlength:int -> ?fresh_buf_size:int -> unit -> t

create ?maxlength ?fresh_buf_size () creates a buffer of size maxlength (by default 32 kb). If the buffer is full, a buffer of size fresh_buf_size is allocated (by default 2 kb).

val write : maxlen:int -> fill_using: (Stdlib.Bytes.t -> int -> int -> (int, 'error) Stdlib.result Lwt.t) -> t -> (data, 'error) Stdlib.result Lwt.t

write ~maxlen ~fill_using buffer calls fill_using buf offset maxlen where buf is a buffer that has room for maxlen data starting from offset.

Assumes that fill_using returns the exact amount of written bytes.

Behaviour is unspecified if fill_using writes more than maxlen data or lies on the number of written bytes.

It returns a data descriptor for the supposedly written chunk.

val read : data -> ?len:int -> t -> into:Stdlib.Bytes.t -> offset:int -> data option

read data ~len ~into:buf buffer ~offset copies len data from the data chunk into buf. If len is not provided, it copies all the data. If len is less than the amount of data available, it returns a new handler of the remainder.

Assumes that data has been produced by a write attempt in buffer.

Assumes that len is less than length data.

val length : data -> int

length data returns the amount of available bytes in data