Module Tezos_stdlib.Hash_queue

Bounded queues combined with hash-tables, based on Aches.

A hash queue is a structure where one can add elements to the back of the queue, while associating them to keys. This allows for efficiently retrieving elements based on the key and for removing elements anywhere in the queue.

module Make (K : Stdlib.Hashtbl.HashedType) (V : sig ... end) : sig ... end