Module Tezos_stdlib.Lwt_pipe

exception Closed

module Bounded : sig ... end

Data queues similar to the Pipe module in Jane Street's Async library. They are implemented with Queues, limited in size, and use Lwt primitives for concurrent access.

module Unbounded : sig ... end

Unbounded is a variant of Bounded where there is no size limit. It is equivalent to using Bounded with compute_size = Fun.const 0 but some functions are specialised to this particular setup.

module Maybe_bounded : sig ... end

Maybe_bounded are pipes that may or may not be bounded. This is decided when createing the pipe and can be queried with the bounded function.