Module Tezos_stdlib.Utils

module Infix : sig ... end

val cut : ?copy:bool -> int -> Stdlib.Bytes.t -> Stdlib.Bytes.t list

cut ?copy size bytes cut bytes the in a list of successive chunks of length size at most.

If copy is false (default), the blocks of the list can be garbage-collected only when all the blocks are unreachable (because of the 'optimized' implementation of sub used internally.

val do_n_times : int -> (unit -> unit) -> unit

do_n_times n f executes f n times. If n is negative, invalid_arg "do_n_times" is executed.

val fold_n_times : int -> ('a -> 'a) -> 'a -> 'a

fold_n_times n f composes f n times. If n is negative, invalid_arg "fold_n_times" is executed.