Module Key_value_store.Internal_for_tests

val init : ?lockfile_prefix:string -> lru_size:int -> root_dir:string -> unit -> ('file, 'key, 'value) t Tezos_error_monad.Error_monad.tzresult Lwt.t

Same as init above, except that the user can specify a prefix for the lock file (default is lockfile_prefix = "internal_for_tests") to avoid issues if the store is already locked by another process, such as:

IO error in lockf(): Resource temporarily unavailable)