Module Key_value_store.View

val opened_files : ('file, 'key, 'value) t -> int

Returns the number of files currently opened by the key value store. Do note this number is an upper bound on the number of file descriptors opened. This number should always be lower than lru_size.

val ongoing_actions : ('file, 'key, 'value) t -> int

Returns the number of ongoing actions happening on different files. This number should always be lower than lru_size.