Module Tezos_stdlib_unix.Level_config_rules

This module handles set of rules for log sections specific configurations. For instance, if one want enable section "node.context" at level debug, it would be written "node.context -> debug". The complete syntax is the following

<rules> := <rule> ; <rules> <rule> := (<pattern> ->)? <level> <pattern> := <string> | * | <pattern>.<pattern>

A value of this form can be found in environment variables TEZOS_LOG and LWT_LOG.

exception Incorrect_log_rules_syntax

exception Incorrect_log_rules_not_a_level of string

exception Incorrect_log_rules_missing_pattern

exception Incorrect_log_rules_missing_level

type rules = (string * Tezos_event_logging.Internal_event.Level.t option) list

val find_log_rules : string option -> string * string option

find_log_rules default checks above environment variables for set of rules

returns the origin (environment variable or config file) and value if found

returns the default given configuration otherwise

val parse_rules : string -> rules

parse_rules s parses the string s representing a set of rules