Module Tezos_test_helpers.Lwt_assert

lwt_assert contains Alcotest assertions lifted to the Lwt.

val lwt_assert_true : string -> bool -> unit Lwt.t

Lwt-lifted assertion that b is true.

val lwt_assert_false : string -> bool -> unit Lwt.t

Lwt-lifted assertion that b is false.

val lwt_check : 'a Alcotest.testable -> string -> 'a -> 'a -> unit Lwt.t

Lwt-lift of Alcotest.check.

val lwt_impossible : string -> unit Lwt.t

Lwt-lifted Alcotest version of assert false.

val lwt_fail : string -> 'a Lwt.t

lwt_fail str is a lwt-lifted Alcotest version of Alcotest.fail str.

val lwt_assert_catch : (unit -> unit Lwt.t) -> exn -> 'a Lwt.t

lwt_assert_catch p e runs p, expecting to catch e and fails if p does not raise that exception