Module Qcheck2_helpers.MakeMapGen

Map-related generators.

Parameters

module Map : sig ... end

Signature

val gen_of_size : int QCheck2.Gen.t -> Map.key QCheck2.Gen.t -> 'v QCheck2.Gen.t -> 'v Map.t QCheck2.Gen.t

gen_of_size size_gen key_gen val_gen is a generator of Map where the keys are generated with key_gen and the values with val_gen. The number of entries in the map is decided by size_gen.

val gen : Map.key QCheck2.Gen.t -> 'v QCheck2.Gen.t -> 'v Map.t QCheck2.Gen.t

gen key_gen gen_gen is a generator of Map where the keys are generated with key_gen and the values with val_gen.