Module Tezos_test_helpers.Qcheck_extra

This library contains unofficial extensions to the QCheck library.

QCheck2_extra.Stateful.Gen works much like QCheck2.Gen, but allows embedding of side-effects.

module Non_empty : sig ... end

A non-empty list

module Functor : sig ... end

Functor abstracts over type constructors that can be mapped over.

module Applicative : sig ... end

Applicative extends Functor with an map2 and return method, allowing us to lift functions of arbitrary arity.

module Monad : sig ... end

Monad extends the Applicative type class with a new function join. join takes a value in a nested context a t t and joins them together so that we have a single context a t.

module Identity : sig ... end

The identity type constructor.

module Stateful_gen : sig ... end