Module Tezos_test_helpers.Tag

Test tags with specific semantics.

The following values are string constants that can be added to ~tags in Test.register. They have specific semantics that is documented here. By using Tag.x instead of "x" in ~tags, you allow developers to jump to the definition of Tag.x and thus see the documentation of the tag. You also avoid mistakes like using "ci_disable" instead of "ci_disabled".

val flaky : string

"flaky": the test is flaky.

The semantics depends on other tags:

if the test also has tag memory_3k, memory_4k or time_sensitive, the test still runs in job tezt-memory-3k, tezt-memory-4k or tezt-time-sensitive respectively;

if the test also has tag slow, the test still runs in job tezt-slow but this job is manual, so it does not run by default, except in scheduled pipelines;

if the test has none of these other tags, the test still runs in job tezt-flaky, but this job is manual, so it does not run by default, except in scheduled pipelines; additionally, this job is allowed to fail.

In the general case, this tag thus disables the test in before_merging pipelines of the CI just like ci_disabled. Contrary to ci_disabled however, this also gives the reason why the test is disabled: the test is flaky.

Tip: you can list tests that are declared as flaky with tezt --list flaky. You can check if they are still flaky with something like tezt flaky --loop-count 100.

Since this tag already conveys information about why the test is disabled, and since it is easy to get the list of tests with this tag, it is not necessarily needed to create an issue for each flaky test.

val ci_disabled : string

"ci_disabled": disable test in the CI.

The test is not to be run in the CI. You must provide a comment to explain why the test is disabled. For flaky tests, flaky should be preferred to ci_disabled.

val memory_3k : string

"memory_3k": tag memory hungry tests ( >3 GB of memory ).

val memory_4k : string

"memory_4k": tag memory hungry tests ( >4 GB of memory ).

val time_sensitive : string

"time_sensitive": tag for time-sensitive tests.

They are executed with -j 1 to ensure that other tests do not affect their executions. However, they are not particularly cpu/memory-intensive hence they do not need to run on a particular machine contrary to performance regression tests.

val infrastructure : string

"infrastructure": tag for tests owned by the Infrastructure product unit.

val layer1 : string

"layer1": tag for tests owned by the Layer 1 product unit.

val tezos2 : string

"tezos2": tag for tests owned by the Tezos 2 product unit.

val etherlink : string

"etherlink": tag for tests owned by the Etherlink product unit.

val base : string

"base": tag for tests related to lib_base.

val shell : string

"shell": tag for tests related to lib_shell and lib_shell_services.

val encodings : string

"encodings": tag for tests related to encodings.

val services : string

"services": tag for tests related to services.

val unix : string

"unix": tag for tests related to unix.

val slow : string

"slow": tag for tests that are too slow for the CI's before_merging pipelines and that will only be run on the scheduled pipeline.

This tag should be used for tests that take more than 2 minutes in the CI.

val cloud : string

cloud: tag for tests that depends on the tezt-cloud library. Such a test uses VMs deployed onto the cloud.