Module type Encodings_util.Bare_tezos_context_sig

type t

type tree

type index

module Tree : Tezos_context_sigs.Context.TREE with type t := t and type key := string list and type value := bytes and type tree := tree

Tree provides immutable, in-memory partial mirror of the context, with lazy reads and delayed writes. The trees are Merkle trees that carry the same hash as the part of the context they mirror.

val init : ?patch_context:(t -> t Tezos_base.TzPervasives.tzresult Lwt.t) -> ?readonly:bool -> ?index_log_size:int -> string -> index Lwt.t

val empty : index -> t