Module Tezos_version.Network_version

type t = {

chain_name : Distributed_db_version.Name.t;

distributed_db_version : Distributed_db_version.t;

p2p_version : P2p_version.t;

}

val pp : Stdlib.Format.formatter -> t -> unit

val encoding : t Data_encoding.t

val announced : chain_name:Distributed_db_version.Name.t -> distributed_db_versions:Distributed_db_version.t list -> p2p_versions:P2p_version.t list -> t

Get the network protocol version to announce on peer connection.

Use the highest distributed_db_versions and the highest p2p_versions. The version also contains the chain_name since it is used to prevent peers from different networks to communicate.

Neither distributed_db_versions nor p2p_versions can be empty.