Module Tezos_webassembly_interpreter

module Ast : sig ... end

module Binary_exn : sig ... end

module Decode : sig ... end

module Durable_storage : sig ... end

module Encode : sig ... end

module Env : sig ... end

module Error : sig ... end

module Eval : sig ... end

module Eval_num : sig ... end

module Eval_vec : sig ... end

module F32 : sig ... end

module F32_convert : sig ... end

module F64 : sig ... end

module F64_convert : sig ... end

module Free : sig ... end

module Func : sig ... end

module Fxx : sig ... end

module Global : sig ... end

module Host_funcs : sig ... end

module I16 : sig ... end

module I32 : sig ... end

module I32_convert : sig ... end

module I64 : sig ... end

module I64_convert : sig ... end

module I8 : sig ... end

module Import : sig ... end

module Input_buffer : sig ... end

This module implements a FIFO queue to model the input. The messages are queued in an input_buffer in their order of appearance in the inbox.

module Instance : sig ... end

module Ixx : sig ... end

module Lib : sig ... end

module Memory : sig ... end

module Memory_sig : sig ... end

module Operators : sig ... end

module Output_buffer : sig ... end

This module implements a bounded FIFO queue to model the outputs. Each element of the queue is a list of outbox messages produced on this level.

module Partial_memory : sig ... end

module Partial_table : sig ... end

module Proto_compat : sig ... end

module Source : sig ... end

module Spectest : sig ... end

module Table : sig ... end

module Table_sig : sig ... end

module Tick_model : sig ... end

This module implements a generic tick model. These values were benchmarked from `MemoryCopy`, `MemoryFill` and `MemoryInit`.

module Types : sig ... end

module Utf8 : sig ... end

module V128 : sig ... end

module Valid : sig ... end

module Values : sig ... end