Module type Table_sig.S

type table

type t = table

type size = int32

type index = int32

type count = int32

exception Type

exception Bounds

exception SizeOverflow

exception SizeLimit

exception OutOfMemory

val alloc : Types.table_type -> Values.ref_ -> table

val type_of : table -> Types.table_type

val size : table -> size

val grow : table -> size -> Values.ref_ -> unit

val load : table -> index -> Values.ref_ Lwt.t

val store : table -> index -> Values.ref_ -> unit

val blit : table -> index -> Values.ref_ list -> unit

val alloc_shallow : Types.table_type -> t

alloc_shallow table_ty is like alloc but without a default value for the references in the table. That means accessing any unset element will raise Lazy_map.UnexpectedAccess.

module Vector = Tezos_lazy_containers.Lazy_vector.Mutable.Int32Vector

val of_lazy_vector : Types.table_type -> Values.ref_ Vector.Vector.t -> t