Module T.Dropbox

include BOX with type t := dropbox t

val put_request : dropbox t -> ('a, 'request_error) Request.t -> unit

put_request worker request sends the request to the worker. If the worker dropbox is closed, then it is a no-op.

val put_request_and_wait : dropbox t -> ('a, 'request_error) Request.t -> ('a, 'request_error message_error) Stdlib.result Lwt.t

put_request_and_wait worker request sends the request to the worker and waits for its completion. If the worker dropbox is closed, then it returns Error Closed.