Module Structs.Make

Make is a functor that takes a Trace as argument and instantiates all the Traced modules based on it.

Parameters

module Trace : Traced_sigs.Trace.S

Signature

module Monad : sig ... end

module Seq : sig ... end

module Hashtbl : sig ... end

module List : sig ... end

module Map : sig ... end

module Option = Bare_structs.Option

module Result = Bare_structs.Result

module Set : sig ... end

module Seq_e = Seq_e

module Seq_s : sig ... end

module Seq_es : sig ... end

module Unit = Bare_structs.Unit

module WithExceptions = Bare_structs.WithExceptions