Module Index_intf.Private_types

type merge_stages = [

| `After

| `After_clear

| `After_first_entry

| `Before

]

Some operations that trigger a merge can have hooks inserted at the following stages:

`Before: immediately before merging (while holding the merge lock);

`After_clear: immediately after clearing the log, at the end of a merge;

`After_first_entry: immediately after adding the first entry in the merge file, if the data file contains at least one entry;

`After: immediately after merging (while holding the merge lock).

type merge_result = [

| `Completed

| `Aborted

]