Module Resto_cohttp_self_serving_client.Self_serving_client

module Make (Encoding : Resto.ENCODING) (Log : Resto_cohttp_server.Server.LOGGING) : sig ... end

Make(Encoding) returns a module that provides an instance of Client.CALL in its launch function. The Client.CALL instance is meant to be passed to Client.Make to obtain a local, self-serving client: a client that is their own server.