Module Server.Make_selfserver

Make_selfserver is a functor that produces only the machinery necessary for local use. Specifically, it produces the values and types needed for the Self_serving_client.

Parameters

module Encoding : Resto.ENCODING

module Log : LOGGING

Signature

module Media_type : module type of struct include Resto_cohttp.Media_type.Make(Encoding) end

module Directory : module type of struct include Resto_directory.Make(Encoding) end

module Media : sig ... end

module Agent : sig ... end

module Handlers : sig ... end