Module Raw.Viewing_key

type ak = R.ak

Public signature key. Note: if this key is exposed there is privacy loss! Only randomised version of it (rk) are published.

type nk = R.nk

Public nullifier key. Note: if this key is exposed there is privacy loss! Only randomised version of it (rk) are published.

type ivk = R.ivk

type pkd = R.pkd

type ovk = R.ovk

type diversifier = R.diversifier

Used to create an address from a viewing key.

val diversifier_encoding : R.diversifier Data_encoding.encoding

type full_viewing_key = R.full_viewing_key = {

ak : ak;

nk : nk;

ovk : ovk;

}

Full viewing key contains ak, nsk, ovk

val fvk_encoding : full_viewing_key Data_encoding.encoding

val of_expsk : Spending_key.expanded_spending_key -> full_viewing_key

type t = R.zip32_full_viewing_key = {

depth : Stdlib.Bytes.t;

parent_fvk_tag : Stdlib.Bytes.t;

child_index : Stdlib.Bytes.t;

chain_code : Stdlib.Bytes.t;

fvk : full_viewing_key;

dk : Stdlib.Bytes.t;

}

Type t additionally contains zip-32 related info

val ovk_of_xfvk : t -> ovk

val encoding : t Data_encoding.t

val to_bytes : R.zip32_full_viewing_key -> Stdlib.Bytes.t

val of_bytes : Stdlib.Bytes.t -> R.zip32_full_viewing_key option

val of_sk : Spending_key.t -> t

type index = R.diversifier_index

val compare_index : R.diversifier_index -> R.diversifier_index -> int

val index_to_int64 : R.diversifier_index -> int64

val index_of_int64 : int64 -> R.diversifier_index

val index_encoding : R.diversifier_index Data_encoding.encoding

val default_index : R.diversifier_index

val index_succ : R.diversifier_index -> R.diversifier_index

type address = {

diversifier : diversifier;

pkd : pkd;

}

val address_encoding : address Data_encoding.encoding

type Tezos_crypto.Base58.data +=

| Data of address

val address_b58check_encoding : address Tezos_crypto.Base58.encoding

val new_address : R.zip32_full_viewing_key -> R.diversifier_index -> R.diversifier_index * address

val to_ivk : t -> R.ivk

val dummy_address : unit -> address